api/requirements.txt
# api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
tiktoken==0.5.2  # For token counting
click==8.1.7
boto3==1.34.7
